fix atoi overflow

+++ libft/libft.h
@@ -15,7 +15,7 @@
 
 # include <stdlib.h>
 # include <unistd.h>
-
+# include <limits.h>
 typedef struct s_stack
 {
 	int index;
@@ -45,7 +45,7 @@ void				*ft_memchr(const void *s, int c, size_t n);
 int					ft_memcmp(const void *s1, const void *s2, size_t n);
 char				*ft_strnstr(const char *haystack, const char *needle,
 						size_t n);
-int					ft_atoi(const char *str);
+long long			ft_atoi(const char *str);
 char				*ft_strdup(const char *s);
 void				*ft_calloc(size_t num_elements, size_t element_size);
 char				*ft_substr(char const *s, unsigned int start, size_t len);
